Moved constants. Put the dummy response feature in its own file
