+++ Hackney.Core/Middleware/HeaderConstants.cs
@@ -1,6 +1,6 @@
-﻿namespace Hackney.Core
+﻿namespace Hackney.Core.Middleware
 {
-    public static class Constants
+    public static class HeaderConstants
     {
         public const string CorrelationId = "x-correlation-id";
         public const string UserId = "x-user-id";
